Все! Разобрался с распределнеиями! Решил строить распределение по периодам так естественнее и симметричнее! --Нехватка событий на больших чатотах заключалась в том, что в неравентстеве для определениея полоежения события в распределении, нужно было использовать для одной границы знак <= т.к. событий у которых занчение равно границы становится больше. --В функции  для построения распределения периодов стал использовать даблы.

+++ cprocessrawepigraphdata.h
@@ -75,12 +75,12 @@ private:
                            DWORD valueFin,
                            const WORD channelNumber,
                            DWORD * destribArray);
-    void PutEventInDestrib(DWORD value,
-                           DWORD valueIn,
-                           DWORD valueFin,
+    void PutEventInDestrib(double value,
+                           double valueIn,
+                           double valueFin,
                            const WORD channelNumber,
                            DWORD * destribArray,
-                           DWORD& );
+                           double&);
     void PutEventInFreqDestrib(double,
                            double,
                            double,
